% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pantry.R
\name{copy_to.src_postgres}
\alias{copy_to.src_postgres}
\title{this adds the fast argument working around a known problem in dplyr and DBI
https://github.com/hadley/dplyr/blob/master/R/tbl-sql.r#L294}
\usage{
copy_to.src_postgres(dest, df, name = deparse(substitute(df)), types = NULL,
  temporary = TRUE, unique_indexes = NULL, indexes = NULL,
  analyze = TRUE, fast = FALSE, ...)
}
\description{
https://github.com/hadley/dplyr/issues/1471
https://github.com/rstats-db/DBI/issues/62
}

