% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{date_code}
\alias{date_code}
\title{create date code of the form YYMMDD
inputs:
   d (optional): date code in the format of Sys.Date() for which to generate the date code, defaulting to 'today'}
\usage{
date_code(d = NA)
}
\description{
create date code of the form YYMMDD
inputs:
   d (optional): date code in the format of Sys.Date() for which to generate the date code, defaulting to 'today'
}
