% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_repo.R
\name{get_data_repo}
\alias{get_data_repo}
\title{attach to the data repository, optionally using a schema. If the
specified schema doesn't exist, create it.}
\usage{
get_data_repo(schema = NULL, login = "~/.data_repo_login")
}
\arguments{
\item{login}{either the name of a json file with login information or a
list of login info. The keys should correspond with the argument
keys of the dplyr::src_postgres function.}
}
\value{
a connection to the the data_repo
}
\description{
attach to the data repository, optionally using a schema. If the
specified schema doesn't exist, create it.
}

