% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pantry.R
\name{get_pantry}
\alias{get_pantry}
\title{attach to the data repository, optionally using a schema. If the
specified schema doesn't exist, create it.}
\usage{
get_pantry(schema = NULL, login_info = "~/.pantry_login")
}
\arguments{
\item{login_info}{see get_pantry_login}
}
\value{
a connection to the the pantry
}
\description{
attach to the data repository, optionally using a schema. If the
specified schema doesn't exist, create it.
}

