% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pantry.R
\name{get_pantry_config}
\alias{get_pantry_config}
\title{Get and check login information for Postgres database}
\usage{
get_pantry_config(pantry_config = "~/.pantry_config")
}
\arguments{
\item{pantry_config}{either the name of a json file with config information or
the config info parsed into a list. The login key should correspond with the argument
keys of the dplyr::src_postgres function.}
}
\value{
list of login parameters for Postgres database
}
\description{
Get and check login information for Postgres database
}
