% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pantry.R
\name{get_pantry_login}
\alias{get_pantry_login}
\title{Get and check login information for Postgres database}
\usage{
get_pantry_login(login_info = "~/.pantry_login")
}
\arguments{
\item{login_info}{either the name of a json file with login information or a
list of login info. The keys should correspond with the argument
keys of the dplyr::src_postgres function.}
}
\value{
list of login parameters for Postgres database
}
\description{
Get and check login information for Postgres database
}

