% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_tests.R
\name{test_distinct}
\alias{test_distinct}
\title{test distinct rows
e.g. for testing for network loops}
\usage{
test_distinct(data, ids)
}
\description{
test distinct rows
e.g. for testing for network loops
}

