% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_tests.R
\name{test_is_symmetric}
\alias{test_is_symmetric}
\title{test if the columns names(ids) are symmetric to the columns ids
and for those that are symmetric, test equality over the value columns}
\usage{
test_is_symmetric(data, ids, values)
}
\description{
test if the columns names(ids) are symmetric to the columns ids
and for those that are symmetric, test equality over the value columns
}

