% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{uniprot_entry_web_lookup}
\alias{uniprot_entry_web_lookup}
\title{Get lookup info for given uniprot entries one at a time}
\usage{
uniprot_entry_web_lookup(
  uniprot_entries,
  columns = c("entry name"),
  format = "tab",
  verbose = FALSE
)
}
\description{
#http://www.uniprot.org/help/uniprotkb_column_names
 given a vector of uniprot entries return information about them
}
