% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{uniprot_taxa}
\alias{uniprot_taxa}
\title{Get all taxa below a given ancestor taxa}
\usage{
uniprot_taxa(ancestor = 7711, reviewed = TRUE, format = "tab",
  verbose = F)
}
\description{
Get all taxa below a given ancestor taxa
}
