% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{uniprot_web_lookup}
\alias{uniprot_web_lookup}
\title{Get lookup info for given uniprot accessions one at a time}
\usage{
uniprot_web_lookup(uniprot_accns, columns = c("entry name"), verbose = F)
}
\description{
Get lookup info for given uniprot accessions one at a time
}
