% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pantry.R
\name{get_pantry}
\alias{get_pantry}
\title{Attach to the pantry database, optionally using a schema. If the
specified schema doesn't exist, create it.}
\usage{
get_pantry(schema = NULL, pantry_config = "~/.pantry_config")
}
\arguments{
\item{pantry_config}{see get_pantry_config}
}
\value{
a connection to the the pantry
}
\description{
Attach to the pantry database, optionally using a schema. If the
specified schema doesn't exist, create it.
}
