% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_repo.R
\name{get_tables}
\alias{get_tables}
\title{Get all tables in the active schema or search path}
\usage{
get_tables(data_repo, all_tables = FALSE)
}
\description{
Get all tables in the active schema or search path
}

