% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_repo.R
\name{schema_tbl}
\alias{schema_tbl}
\title{Identify a table qualified by a schema. This works around issue 244 in dplyr}
\usage{
schema_tbl(data_repo, schema_table)
}
\description{
Identify a table qualified by a schema. This works around issue 244 in dplyr
}

